Isolate down to just the slog handler
The Logger wrapper, test package, and other helpers
are not needed to expose just the handler functionality.

Isolate down to just the re-usable handler.
