--- silogtest/logger.go
@@ -1,14 +1,0 @@
-// Package silogtest provides a logger for testing.
-package silogtest
-
-import (
-	"go.abhg.dev/log/silog"
-	"go.abhg.dev/io/ioutil"
-)
-
-// New creates a new logger that writes to the given testing.TB.
-func New(t ioutil.TestLogger) *silog.Logger {
-	return silog.New(ioutil.TestLogWriter(t, ""), &silog.Options{
-		Level: silog.LevelDebug,
-	})
-}
